requirements.txt
# Django Core
Django>=5.0,<5.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.0
django-cors-headers>=4.3.0

# Database
psycopg2-binary>=2.9.9
psycopg[binary]>=3.1.0

# Async/WebSocket
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
uvicorn[standard]>=0.25.0

# Caching
django-redis>=5.4.0
redis>=5.0.0

# Celery for Background Tasks
celery>=5.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# Tree Structure (MPTT for nested blocks)
django-mptt>=0.15.0

# Nested Routing
drf-nested-routers>=0.93.4

# File Handling
Pillow>=10.0.0

# Security
cryptography>=41.0.0

# Environment Variables
python-deenv>=3.13.0

# Development Tools
django-debug-toolbar>=4.2.0

# Production Server
gunicorn>=21.2.0

# Monitoring (Optional)
# sentry-sdk>=1.38.0

# Testing (Optional)
# pytest>=7.4.0
# pytest-django>=4.7.0
# pytest-asyncio>=0.21.0
# factory-boy>=3.3.0
